requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyMySQL==1.1.1
Werkzeug==3.0.3
pandas==2.2.2
python-dotenv==1.0.1
gunicorn==22.0.0
